First Updates for Nano AOD

+++ Tools/scripts/getStopCfg.sh
@@ -14,6 +14,7 @@ RELEASE_URL="$GITHUB_SUSY2015_URL/$REPO_NAME/releases"
 STARTING_DIR=$PWD
 CFG_DIRECTORY=$PWD
 TAG=
+USE_PRE=
 NO_SOFTLINK=
 OVERWRITE=
 VERBOSE=
@@ -22,10 +23,18 @@ VERBOSE=
 OPTIND=1    # Reset in case getopts has been used previously in the shell.
 
 SUPP_CFG=supplementaryFiles.cfg
-SETS_CFG=sampleSets.cfg
-COLL_CFG=sampleCollections.cfg
-SETS_LINK_NAME=$SETS_CFG
-COLL_LINK_NAME=$COLL_CFG
+SETS_CFG_2016=sampleSets_PostProcessed_2016.cfg
+SETS_CFG_2017=sampleSets_PostProcessed_2017.cfg
+SETS_CFG_2018=sampleSets_PostProcessed_2018.cfg
+COLL_CFG_2016=sampleCollections_2016.cfg
+COLL_CFG_2017=sampleCollections_2017.cfg
+COLL_CFG_2018=sampleCollections_2018.cfg
+SETS_LINK_NAME_2016=$SETS_CFG_2016
+SETS_LINK_NAME_2017=$SETS_CFG_2017
+SETS_LINK_NAME_2018=$SETS_CFG_2018
+COLL_LINK_NAME_2016=$COLL_CFG_2016
+COLL_LINK_NAME_2017=$COLL_CFG_2017
+COLL_LINK_NAME_2018=$COLL_CFG_2018
 SUPP_FILE_DIR=supplementaryFiles
 TARBALL=SUPP_FILE_TARBALL.tar.gz
 
@@ -37,10 +46,15 @@ function print_help {
     echo "Options:"
     echo "    -t RELEASE_TAG :         This is the github release tag to check out (required option)"
     echo "    -d checkout_directory :  This is the directory where the configuration files will be downloaded to (default: .)"
-    echo "    -s SETS_LINK_NAME :      Specify this option to name the softlink to \"$SETS_CFG\" something other than \"$SETS_LINK_NAME\""
-    echo "    -c COLL_LINK_NAME :      Specify this option to name the softlink to \"$COLL_CFG\" something other than \"$COLL_LINK_NAME\""
-    echo "    -o :                     Overwrite the softlinks if they already exist"
+    echo "    -a SETS_LINK_NAME_2016 : Specify this option to name the softlink to \"$SETS_CFG_2016\" something other than \"$SETS_LINK_NAME_2016\""
+    echo "    -b SETS_LINK_NAME_2017 : Specify this option to name the softlink to \"$SETS_CFG_2017\" something other than \"$SETS_LINK_NAME_2017\""
+    echo "    -c SETS_LINK_NAME_2018 : Specify this option to name the softlink to \"$SETS_CFG_2018\" something other than \"$SETS_LINK_NAME_2018\""
+    echo "    -x COLL_LINK_NAME_2016 : Specify this option to name the softlink to \"$COLL_CFG_2016\" something other than \"$COLL_LINK_NAME_2016\""
+    echo "    -y COLL_LINK_NAME_2017 : Specify this option to name the softlink to \"$COLL_CFG_2017\" something other than \"$COLL_LINK_NAME_2017\""
+    echo "    -z COLL_LINK_NAME_2018 : Specify this option to name the softlink to \"$COLL_CFG_2018\" something other than \"$COLL_LINK_NAME_2018\""
+    echo "    -e :                     Softlink pre-processed config files."
     echo "    -n :                     Download files without producing softlinks"
+    echo "    -o :                     Overwrite the softlinks if they already exist"
     echo "    -v :                     increase verbosity: print more stuff... for those who like stuff"
     echo ""
     echo "Description:"
@@ -54,7 +68,7 @@ function print_help {
 
 # Initialize our own variables:
 
-while getopts "h?t:d:s:c:onv" opt; do
+while getopts "h?t:d:a:b:x:y:enov" opt; do
     case "$opt" in
     h|\?)
         print_help
@@ -64,13 +78,23 @@ while getopts "h?t:d:s:c:onv" opt; do
         ;;
     d)  CFG_DIRECTORY=$OPTARG
         ;;
-    s)  SETS_LINK_NAME=$OPTARG
+    a)  SETS_LINK_NAME_2016=$OPTARG
         ;;
-    c)  COLL_LINK_NAME=$OPTARG
+    b)  SETS_LINK_NAME_2017=$OPTARG
         ;;
-    o) OVERWRITE="-f"
+    c)  SETS_LINK_NAME_2018=$OPTARG
+        ;;
+    x)  COLL_LINK_NAME_2016=$OPTARG
+        ;;
+    y)  COLL_LINK_NAME_2017=$OPTARG
+        ;;
+    z)  COLL_LINK_NAME_2018=$OPTARG
+        ;;
+    e) USE_PRE=1
         ;;
     n) NO_SOFTLINK=1
+        ;;
+    o) OVERWRITE="-f"
         ;;
     v) VERBOSE=1
         ;;
@@ -88,6 +112,17 @@ then
 fi
 
 echo " - Running getStopCfg.sh"
+
+# if USE_PRE is set, softlink pre-processed configs
+if [[ ! -z $USE_PRE ]]
+then
+    SETS_CFG_2016=sampleSets_PreProcessed_2016.cfg
+    SETS_CFG_2017=sampleSets_PreProcessed_2017.cfg
+    SETS_CFG_2018=sampleSets_PreProcessed_2018.cfg
+    SETS_LINK_NAME_2016=$SETS_CFG_2016
+    SETS_LINK_NAME_2017=$SETS_CFG_2017
+    SETS_LINK_NAME_2018=$SETS_CFG_2018
+fi
 
 # get source directory of bash script
 # used for "Easter Egg"...
@@ -224,8 +259,12 @@ if [[ -z $NO_SOFTLINK ]]
 then
     # create softlinks
     echo " - Creating softlinks to $REPO_NAME config files"
-    ln $OVERWRITE -s $DOWNLOAD_DIR/$SETS_CFG $SETS_LINK_NAME > /dev/null 2>&1 
-    ln $OVERWRITE -s $DOWNLOAD_DIR/$COLL_CFG $COLL_LINK_NAME > /dev/null 2>&1
+    ln $OVERWRITE -s $DOWNLOAD_DIR/$SETS_CFG_2016 $SETS_LINK_NAME_2016 > /dev/null 2>&1 
+    ln $OVERWRITE -s $DOWNLOAD_DIR/$SETS_CFG_2017 $SETS_LINK_NAME_2017 > /dev/null 2>&1 
+    ln $OVERWRITE -s $DOWNLOAD_DIR/$SETS_CFG_2018 $SETS_LINK_NAME_2018 > /dev/null 2>&1 
+    ln $OVERWRITE -s $DOWNLOAD_DIR/$COLL_CFG_2016 $COLL_LINK_NAME_2016 > /dev/null 2>&1
+    ln $OVERWRITE -s $DOWNLOAD_DIR/$COLL_CFG_2017 $COLL_LINK_NAME_2017 > /dev/null 2>&1
+    ln $OVERWRITE -s $DOWNLOAD_DIR/$COLL_CFG_2018 $COLL_LINK_NAME_2018 > /dev/null 2>&1
     # make list of files to tar
     if [[ -f $DOWNLOAD_DIR/$SUPP_CFG ]]
     then
@@ -233,7 +272,10 @@ then
         while read -r line || [[ -n "$line" ]]
         do
             SUPP_FILE="$line"
-            ln $OVERWRITE -s $DOWNLOAD_DIR/$SUPP_FILE_DIR/$SUPP_FILE $SUPP_FILE > /dev/null 2>&1
+            if [ -f $DOWNLOAD_DIR/$SUPP_FILE_DIR/$SUPP_FILE ]
+            then
+                ln $OVERWRITE -s $DOWNLOAD_DIR/$SUPP_FILE_DIR/$SUPP_FILE $SUPP_FILE > /dev/null 2>&1
+            fi
         done < $DOWNLOAD_DIR/$SUPP_CFG
     else
         echo "ERROR: Config file $DOWNLOAD_DIR/$SUPP_CFG not found"
